+++ docs/storybook/stories/ui/Heading.stories.tsx
@@ -1,5 +1,7 @@
-import { Flex, Heading } from '@ttoss/ui';
-import { Meta, Story } from '@storybook/react';
+import * as React from 'react';
+import { Box, Flex, Heading } from '@ttoss/ui';
+import { Label, Select, useTheme } from '@ttoss/ui/src';
+import { Meta, StoryFn } from '@storybook/react';
 
 export default {
   title: 'UI/Heading',
@@ -14,7 +16,7 @@ export default {
 
 const as = ['h1', 'h2', 'h3', 'h4', 'h5', 'h6'] as const;
 
-const Template: Story = (args) => {
+const Template: StoryFn = (args) => {
   return (
     <Flex sx={{ flexDirection: 'column', gap: 2 }}>
       {as.map((tag) => {
@@ -27,5 +29,54 @@ const Template: Story = (args) => {
     </Flex>
   );
 };
+
+const TemplateLineHeights: StoryFn = (args) => {
+  const { theme } = useTheme();
+  const lineHeights = Object.keys(
+    theme.lineHeights as unknown as { [key: string]: string }
+  );
+
+  const [selectedLineHeight, setSelectedLineHeight] =
+    React.useState<string>('base');
+
+  return (
+    <Flex sx={{ flexDirection: 'column', gap: '2xl' }}>
+      <Box>
+        <Label>Line Height</Label>
+
+        <Select
+          value={selectedLineHeight}
+          onChange={(e) => {
+            return setSelectedLineHeight(e.target.value);
+          }}
+        >
+          {lineHeights.map((lineHeight) => {
+            return (
+              <option key={lineHeight} value={lineHeight}>
+                {lineHeight}
+              </option>
+            );
+          })}
+        </Select>
+      </Box>
+      <Flex sx={{ flexDirection: 'column', gap: 'lg' }}>
+        {as.map((tag) => {
+          return (
+            <Heading
+              sx={{ lineHeight: selectedLineHeight, backgroundColor: 'muted' }}
+              as={tag}
+              variant={tag}
+              key={tag}
+            >
+              {args.text} {tag}
+            </Heading>
+          );
+        })}
+      </Flex>
+    </Flex>
+  );
+};
+
+export const ExampleWithDifferentLineHeights = TemplateLineHeights.bind({});
 
 export const Example = Template.bind({});