+++ packages/theme/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.3](https://github.com/ttoss/ttoss/compare/@ttoss/theme@1.6.2...@ttoss/theme@1.6.3) (2023-08-23)
+
+**Note:** Version bump only for package @ttoss/theme
+
+## [1.6.2](https://github.com/ttoss/ttoss/compare/@ttoss/theme@1.6.1...@ttoss/theme@1.6.2) (2023-08-22)
+
+**Note:** Version bump only for package @ttoss/theme
+
 ## [1.6.1](https://github.com/ttoss/ttoss/compare/@ttoss/theme@1.6.0...@ttoss/theme@1.6.1) (2023-08-16)
 
 ### Bug Fixes