+++ packages/ui/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.3](https://github.com/ttoss/ttoss/compare/@ttoss/ui@3.1.2...@ttoss/ui@3.1.3) (2023-08-23)
+
+**Note:** Version bump only for package @ttoss/ui
+
+## [3.1.2](https://github.com/ttoss/ttoss/compare/@ttoss/ui@3.1.1...@ttoss/ui@3.1.2) (2023-08-22)
+
+**Note:** Version bump only for package @ttoss/ui
+
 ## [3.1.1](https://github.com/ttoss/ttoss/compare/@ttoss/ui@3.1.0...@ttoss/ui@3.1.1) (2023-08-16)
 
 ### Bug Fixes